refactor(web): authenticateの散在するロジックをlib/に集約
feature直下に散らばっていた error-message / turnstile / validation /
validation-result を lib/ にまとめ、components・hooks と並ぶ構成に整理する。
公開API（index.ts）と参照側の挙動は変えていない。
