+++ application/packages/web/src/client/features/authenticate/components/authenticate-form.tsx
@@ -3,7 +3,7 @@ import { Button } from '@/client/components/shadcn/button'
 import { Input } from '@/client/components/shadcn/input'
 import { Label } from '@/client/components/shadcn/label'
 import { TurnstileWidget } from '@/client/features/authenticate/components/turnstile-widget'
-import type { AuthenticateErrors } from '@/client/features/authenticate/validation'
+import type { AuthenticateErrors } from '@/client/features/authenticate/lib/validation'
 
 interface Props {
   onSubmit: (formData: FormData) => void
